Added the compactor block
It doesn't actually do anything, just preparing it. Faced an annoying
bug that occured any time trying to load the world, it would go back to
the menu but still be consuming cpu. I had added everything to BlockInfo
and the other necessary classes, but forgot to set the ID in the
configHandler. Therefore, it never had an ID, screwing things up.