Attempted to fix some errors... didn't succeed
Added success count to placement util to see if it thought it was
placing the blocks since they never showed up, it did, it probably has
something to do with the itemID it thinks it's placing is 0 or
something... not sure. Also tried to fix logic with determineHomogenous
since it was working incorrectly, didnt though.