Tried to sync players in the beginning when rendering is empty
However, I uncovered a much larger issue that would likely solve the
rendering issue. When the contents of the compactor are modified, the
changes aren't sent to all players that might be using it. This means
that you can duplicate items by you putting the items in and someone
else taking them out. Big problem. Not Today.